+++ src/dashboard/apps/home/apps.py
@@ -1,0 +1,10 @@
+"""Dashboard home app base config."""
+
+from django.apps import AppConfig
+
+
+class HomeConfig(AppConfig):
+    """Home app config."""
+
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "apps.home"