+++ src/dashboard/dashboard/urls.py
@@ -16,8 +16,10 @@ Including another URLconf
 """
 
 from django.contrib import admin
-from django.urls import path
+from django.urls import include, path
 
 urlpatterns = [
     path("admin/", admin.site.urls),
+    path("", include("apps.home.urls")),
+    path("i18n/", include("django.conf.urls.i18n")),
 ]